+++ ghost/core/core/frontend/web/site.js
@@ -2,7 +2,7 @@ const debug = require('@tryghost/debug')('frontend');
 const path = require('path');
 const express = require('../../shared/express');
 const DomainEvents = require('@tryghost/domain-events');
-const {MemberPageViewEvent} = require('@tryghost/member-events');
+const {MemberPageViewEvent} = require('../../shared/events');
 
 // App requires
 const config = require('../../shared/config');
