+++ ghost/core/core/shared/events/MemberCreatedEvent.js
