+++ ghost/core/core/shared/events/MemberEntryViewEvent.js
