+++ ghost/core/core/shared/events/MemberSignupEvent.js
