+++ ghost/core/core/shared/events/MemberSubscribeEvent.js
