+++ ghost/core/core/shared/events/SubscriptionCancelledEvent.js
