+++ ghost/core/core/shared/events/index.js
@@ -1,0 +1,17 @@
+module.exports = {
+    MemberCommentEvent: require('./MemberCommentEvent'),
+    MemberCreatedEvent: require('./MemberCreatedEvent'),
+    MemberEntryViewEvent: require('./MemberEntryViewEvent'),
+    MemberLinkClickEvent: require('./MemberLinkClickEvent'),
+    MemberPageViewEvent: require('./MemberPageViewEvent'),
+    MemberPaidCancellationEvent: require('./MemberPaidCancellationEvent'),
+    MemberPaidConversionEvent: require('./MemberPaidConversionEvent'),
+    MemberSignupEvent: require('./MemberSignupEvent'),
+    MemberSubscribeEvent: require('./MemberSubscribeEvent'),
+    MemberUnsubscribeEvent: require('./MemberUnsubscribeEvent'),
+    OfferRedemptionEvent: require('./OfferRedemptionEvent'),
+    SubscriptionActivatedEvent: require('./SubscriptionActivatedEvent'),
+    SubscriptionCancelledEvent: require('./SubscriptionCancelledEvent'),
+    SubscriptionCreatedEvent: require('./SubscriptionCreatedEvent'),
+    URLResourceUpdatedEvent: require('./URLResourceUpdatedEvent')
+};
