+++ ghost/core/package.json
@@ -101,7 +101,6 @@
     "@tryghost/link-replacer": "0.0.0",
     "@tryghost/logging": "2.4.21",
     "@tryghost/member-attribution": "0.0.0",
-    "@tryghost/member-events": "0.0.0",
     "@tryghost/members-csv": "0.0.0",
     "@tryghost/members-offers": "0.0.0",
     "@tryghost/metrics": "1.0.37",
