+++ ghost/core/test/e2e-frontend/members.test.js
@@ -8,7 +8,7 @@ const configUtils = require('../utils/configUtils');
 const settingsCache = require('../../core/shared/settings-cache');
 const settingsHelpers = require('../../core/server/services/settings-helpers');
 const DomainEvents = require('@tryghost/domain-events');
-const {MemberPageViewEvent} = require('@tryghost/member-events');
+const {MemberPageViewEvent} = require('../../core/shared/events');
 const models = require('../../core/server/models');
 const {fixtureManager} = require('../utils/e2e-framework');
 const DataGenerator = require('../utils/fixtures/data-generator');
