+++ ghost/core/test/unit/server/services/members-events/last-seen-at-updater.test.js
@@ -3,7 +3,7 @@ const assert = require('assert/strict');
 const sinon = require('sinon');
 const LastSeenAtUpdater = require('../../../../../core/server/services/members-events/LastSeenAtUpdater');
 const DomainEvents = require('@tryghost/domain-events');
-const {MemberPageViewEvent, MemberCommentEvent, MemberSubscribeEvent, MemberLinkClickEvent} = require('@tryghost/member-events');
+const {MemberPageViewEvent, MemberCommentEvent, MemberSubscribeEvent, MemberLinkClickEvent} = require('../../../../../core/shared/events');
 const moment = require('moment');
 const {EmailOpenedEvent} = require('@tryghost/email-service');
 const EventEmitter = require('events');
