+++ ghost/member-attribution/package.json
@@ -21,7 +21,6 @@
   },
   "dependencies": {
     "@tryghost/domain-events": "0.0.0",
-    "@tryghost/member-events": "0.0.0",
     "@tryghost/referrer-parser": "0.1.1",
     "@tryghost/string": "0.2.13"
   }
