--- ghost/member-events/index.js
@@ -1,16 +1,0 @@
-module.exports = {
-    MemberCreatedEvent: require('./lib/MemberCreatedEvent'),
-    MemberEntryViewEvent: require('./lib/MemberEntryViewEvent'),
-    MemberSubscribeEvent: require('./lib/MemberSubscribeEvent'),
-    MemberUnsubscribeEvent: require('./lib/MemberUnsubscribeEvent'),
-    MemberSignupEvent: require('./lib/MemberSignupEvent'),
-    MemberPaidConverstionEvent: require('./lib/MemberPaidConversionEvent'),
-    MemberPaidCancellationEvent: require('./lib/MemberPaidCancellationEvent'),
-    MemberPageViewEvent: require('./lib/MemberPageViewEvent'),
-    MemberCommentEvent: require('./lib/MemberCommentEvent'),
-    SubscriptionCreatedEvent: require('./lib/SubscriptionCreatedEvent'),
-    SubscriptionActivatedEvent: require('./lib/SubscriptionActivatedEvent'),
-    SubscriptionCancelledEvent: require('./lib/SubscriptionCancelledEvent'),
-    OfferRedemptionEvent: require('./lib/OfferRedemptionEvent'),
-    MemberLinkClickEvent: require('./lib/MemberLinkClickEvent')
-};
